pyproject.toml
[project]
name = "jtlutil"
version = "0.1.4"
description = "Support routines for League infrastructure"
readme = "README.md"
authors = [
    { name = "Eric Busboom", email = "eric@civicknowledge.com" }
]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-session>=0.8.0",
    "pycryptodome>=3.21.0",
    "pyjwt>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "pytz>=2024.2",
    "requests>=2.32.3",
    "sqlitedict>=2.1.0",

    "docker>=7.1.0",
    "ipykernel>=6.29.5",
    "names-generator>=0.2.0",
    "paramiko>=3.5.0",
    "python-slugify>=8.0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
